refactor: separated form component into hook, context files, and ui
